event_crawler/event_crawler/requirements.txt
scrapy==2.12.0
psycopg2-binary==2.9.6